+++ vrp-pragmatic/src/format/solution/mod.rs
@@ -41,6 +41,9 @@ fn map_code_reason(code: i32) -> (&'static str, &'static str) {
         }
         TOUR_ORDER_CONSTRAINT_CODE => ("TOUR_ORDER_CONSTRAINT_CODE", "cannot be assigned due to tour order constraint"),
         GROUP_CONSTRAINT_CODE => ("GROUP_CONSTRAINT_CODE", "cannot be assigned due to group constraint"),
+        COMPATIBILITY_CONSTRAINT_CODE => {
+            ("COMPATIBILITY_CONSTRAINT_CODE", "cannot be assigned due to compatibility constraint")
+        }
         _ => ("NO_REASON_FOUND", "unknown"),
     }
 }
@@ -61,6 +64,7 @@ fn map_reason_code(reason: &str) -> i32 {
         "TOUR_SIZE_CONSTRAINT" => TOUR_SIZE_CONSTRAINT_CODE,
         "TOUR_ORDER_CONSTRAINT_CODE" => TOUR_ORDER_CONSTRAINT_CODE,
         "GROUP_CONSTRAINT_CODE" => GROUP_CONSTRAINT_CODE,
+        "COMPATIBILITY_CONSTRAINT_CODE" => COMPATIBILITY_CONSTRAINT_CODE,
         _ => -1,
     }
 }