+++ vrp-pragmatic/tests/generator/defaults.rs
@@ -50,6 +50,7 @@ pub fn default_delivery_prototype() -> impl Strategy<Value = Job> {
         generate_no_jobs_skills(),
         generate_no_jobs_value(),
         generate_no_jobs_group(),
+        generate_no_jobs_compatibility(),
     )
 }
 
@@ -59,6 +60,7 @@ pub fn default_pickup_prototype() -> impl Strategy<Value = Job> {
         generate_no_jobs_skills(),
         generate_no_jobs_value(),
         generate_no_jobs_group(),
+        generate_no_jobs_compatibility(),
     )
 }
 
@@ -71,6 +73,7 @@ pub fn default_pickup_delivery_job_prototype() -> impl Strategy<Value = Job> {
         generate_no_jobs_skills(),
         generate_no_jobs_value(),
         generate_no_jobs_group(),
+        generate_no_jobs_compatibility(),
     )
 }